fix: 王的寶箱走到旁邊沒有真的打開 + Gearsetter IPC 型別對不上
BossLoot() 原本只有移動過去、從沒呼叫過互動，寶箱永遠原封不動；
借用既有的 Interactable(ulong?) 走到+互動流程實際把它打開。

GetRecommendationsForGearset 回傳的 tuple 裡 SourceInventorySlot 宣告
成 int?，但 ECommons.IPC 實際回傳的是 byte?，型別對不上導致完全編譯
不過，改回 byte? 讓兩邊一致。

+++ AutoDuty/Helpers/AutoEquipHelper.cs
@@ -112,7 +112,7 @@ namespace AutoDuty.Helpers
             }
         }
 
-        private List<(uint ItemId, InventoryType? SourceInventory, int? SourceInventorySlot, RaptureGearsetModule.GearsetItemIndex TargetSlot)>? _gearset           = null;
+        private List<(uint ItemId, InventoryType? SourceInventory, byte? SourceInventorySlot, RaptureGearsetModule.GearsetItemIndex TargetSlot)>? _gearset           = null;
         private int                                                                                                                               _index             = 0;
 
         private void AutoEquipGearSetterUpdate(IFramework framework)
